// src/service/API/vu-an/types.ts
export const VuDiaDiemResponse_Schema = z.object({
  id: z.string(),
  idDuLieu: z.string(),
  loaiDuLieu: z.string(),
  diaDiemChiTiet: z.any().nullish(),
  xa: z.any().nullish(),
  maXa: z.any().nullish(),
  huyen: z.any().nullish(),
  maHuyen: z.any().nullish(),
  tinh: z.string().nullish(),
  maTinh: z.string().nullish(),
  tuyen: z.any().nullish(),
  maTuyen: z.any().nullish(),
  dacDiemDiaHinh: z.any().nullish(),
  maDacDiemDiaHinh: z.string().nullish(),
  loaiDiaBan: z.any().nullish(),
  maLoaiDiaBan: z.any().nullish(),
  diaBanDacTrung: z.any().nullish(),
  maDiaBanDacTrung: z.string().nullish(),
  xacNhan: z.any().nullish(),
  ngayXayRa: z.any().nullish(),
  gioXayRa: z.any().nullish(),
  xayRaKhoang: z.string().nullish(),
  danhDauXoa: z.any().nullish(),
  isXacDinhDiaDiemCT: z.any().nullish(),
  isDiaDiemChinh: z.boolean().nullish()
});

export const VuDieuLuatResponse_Schema = z.object({
  id: z.string(),
  idDuLieu: z.string(),
  loaiDuLieu: z.string(),
  maDieuLuat: z.string().nullish(),
  tenDieuLuat: z.string().nullish(),
  dieuLuatChinh: z.boolean().nullish(),
  daChuyen: z.boolean().nullish(),
  dieu: z.string().nullish(),
  khoan: z.string().nullish(),
  diem: z.any().nullish(),
  namBoLuat: z.string().nullish(),
  mucDoNghiemTrong: z.string().nullish(),
  khoanTd: z.any().nullish(),
  dieuTd: z.any().nullish(),
  namBoLuatTd: z.any().nullish(),
  phanLoaiToiDanh: z.any().nullish(),
  tenDieuLuatRutGon: z.string().nullish(),
  moTaDieuLuat: z.string().nullish(),
  thuTuMucDoNghiemTrong: z.number().nullish(),
  maPhanLoaiToiDanh: z.any().nullish(),
  tenPhanLoaiToiDanh: z.any().nullish(),
  danhDauTach: z.any().nullish(),
  trangThai: z.number().nullish()
});

export const VuAnThongKeResponse_Schema = z.object({
  id: z.string(),
  idVuAn: z.string().nullish(),
  duocXaHoiQuanTam: z.boolean().nullish(),
  hdBangNhom: z.boolean().nullish(),
  yeuToNuocNgoai: z.boolean().nullish(),
  khongGianMang: z.boolean().nullish(),
  danhSachLuaDao: z.any().nullish(),
  giaTriThietHai: z.any().nullish(),
  maDonViTienTe: z.any().nullish(),
  moTa: z.any().nullish(),
  soNguoiChet: z.number().nullish(),
  soBiThuong: z.number().nullish(),
  soMatTich: z.any().nullish(),
  soTienPhaiThuHoi: z.number().nullish(),
  soTienDaThuHoi: z.number().nullish(),
  thuHoiDu: z.boolean().nullish(),
  loaiKetQua: z.any().nullish(),
  danhDauXoa: z.boolean().nullish(),
  tinDungDen: z.boolean().nullish(),
  donViKetluan: z.any().nullish(),
  tienTePhaiThuHoi: z.any().nullish(),
  tienTeDaThuHoi: z.any().nullish(),
  thongtinThietHai: z.any().nullish(),
  thongtinThuhoi: z.any().nullish(),
  moTaGiaTriThietHai: z.any().nullish(),
  soLuongSung: z.number().nullish(),
  donViSung: z.any().nullish(),
  moTaSoLuongSung: z.any().nullish(),
  soLuongThuocNo: z.number().nullish(),
  donViThuocNo: z.any().nullish(),
  moTaSoLuongThuocNo: z.any().nullish(),
  soLuongTien: z.number().nullish(),
  donViTien: z.any().nullish(),
  moTaSoLuongTien: z.any().nullish(),
  soLuongNgoaiTe: z.any().nullish(),
  donViNgoaiTe: z.any().nullish(),
  moTaSoLuongNgoaiTe: z.any().nullish(),
  soLuongGiayTo: z.number().nullish(),
  giaTriGiayTo: z.number().nullish(),
  donViGiaTriGiayTo: z.any().nullish(),
  moTaGiaTriGiayTo: z.any().nullish(),
  moTaSoLuongGiayTo: z.any().nullish(),
  soLuongBds: z.string().nullish(),
  giaTriBds: z.number().nullish(),
  donViGiaTriBds: z.any().nullish(),
  moTaGiaTriBds: z.any().nullish(),
  moTaSoLuongBds: z.any().nullish(),
  soLuongOto: z.number().nullish(),
  giaTriOto: z.number().nullish(),
  donViGiaTriOto: z.any().nullish(),
  moTaGiaTriOto: z.any().nullish(),
  moTaSoLuongOto: z.any().nullish(),
  giaTriLoaiKhac: z.number().nullish(),
  donViLoaiKhac: z.any().nullish(),
  moTaGiaTriLoaiKhac: z.any().nullish(),
  maDoMat: z.any().nullish(),
  maDonViDongBo: z.any().nullish(),
  vuKhiNong: z.boolean().nullish(),
  soBangNhom: z.any().nullish(),
  soDoiTuongBiBat: z.any().nullish(),
  taiSanThuGiu: z.number().nullish(),
  soLuongHeroin: z.number().nullish(),
  soLuongThuocPhien: z.number().nullish(),
  soLuongCanSa: z.number().nullish(),
  soLuongCayChuaMaTuy: z.number().nullish(),
  khoiLuongMaTuyTongHop: z.number().nullish(),
  soLuongMaTuyTongHop: z.number().nullish(),
  khoiLuongLoaiKhac: z.number().nullish(),
  dungTichLoaiKhac: z.number().nullish(),
  taiSanThietHai: z.number().nullish(),
  thuGiuVietNam: z.number().nullish(),
  thuGiuNgoaiTe: z.number().nullish(),
  banChiDaoTrungUong: z.any().nullish(),
  banChiDaoTinh: z.any().nullish(),
  dieuTraMoRong: z.any().nullish(),
  coToChuc: z.any().nullish(),
  soNanNhanNam: z.number().nullish(),
  soNanNhanNu: z.number().nullish(),
  soPhuNu: z.number().nullish(),
  soTreEmNam: z.number().nullish(),
  soTreEmNu: z.number().nullish(),
  soTroVeTong: z.number().nullish(),
  soTroVeNu: z.number().nullish(),
  soHuongDv: z.number().nullish(),
  soTaiHoaNhapNu: z.number().nullish(),
  so1216Tuoi: z.number().nullish(),
  so1618Tuoi: z.number().nullish(),
  so1860Tuoi: z.number().nullish(),
  soTren60Tuoi: z.number().nullish(),
  soDuoi12Tuoi: z.number().nullish(),
  loaiHiepDam: z.any().nullish(),
  loaiGietNguoi: z.any().nullish(),
  loaiChongNguoiThcv: z.any().nullish(),
  tongTaiSanThietHai: z.number().nullish(),
  taiSanThuHoi: z.number().nullish(),
  taiSanThuHoiKhac: z.number().nullish(),
  taiSanUocTinh: z.number().nullish(),
  soBangNhomTrietPha: z.any().nullish(),
  soDoiTuongBiBatGiu: z.any().nullish(),
  toChucBiThietHai: z.any().nullish(),
  soTienThietHai: z.number().nullish(),
  chongNguoiThiHanhCongVu: z.any().nullish(),
  dauVaoNguonTinC02: z.any().nullish(),
  dauVaoNguonTinC03: z.any().nullish(),
  chuyenDe: z.any().nullish(),
  linhVuc: z.any().nullish(),
  hinhThucToChuc: z.any().nullish(),
  hinhThucTieuThu: z.any().nullish(),
  phuongThucThuDoan: z.any().nullish(),
  diaDiemBatGiu: z.any().nullish(),
  tinhChatVanChuyenMaTuy: z.any().nullish(),
  banNoiChinhTrungUong: z.any().nullish(),
  soDoiTuongSuuTraKhacHe: z.any().nullish(),
  soDoiTuongSuuTraDungHe: z.any().nullish(),
  soDoiTuongNgaoDa: z.any().nullish(),
  soDoiTuongTamThan: z.any().nullish()
});

export interface VuAnResponse extends z.infer<typeof VuAnResponse_Schema> {}
export const VuAnResponse_Schema = z.object({
  id: z.string(),
  maVuAn: z.string().nullish(),
  thoiHan: z.string().nullish(),
  thoiHieu: z.string().nullish(),
  soHoSo: z.string().nullish(),
  maVuViec: z.string().nullish(),
  donVi: z.string().nullish(),
  mucDo: z.string().nullish(),
  maMucDo: z.string().nullish(),
  xacNhanDiaDiem: z.string().nullish(),
  trangThaiVanBan: z.string().nullish(),
  trangThaiVu: z.string().nullish(),
  tenVuAn: z.string().nullish(),
  tenVu: z.string().nullish(),
  noiDung: z.string().nullish(),
  nguonKhoiTo: z.string().nullish(),
  maLoaiNguonVuAn: z.string().nullish(),
  maPhanLoaiDonViKienNghi: z.string().nullish(),
  soVanBanKienNghi: z.string().nullish(),
  ngayKienNghi: z.string().nullish(),
  maPhanLoaiDonViKhoiTo: z.string().nullish(),
  tenPhanLoaiDonViKhoiTo: z.string().nullish(),
  ngayTiepNhanTuTinBao: z.string().nullish(),
  donViPhuTrach: z.string().nullish(),
  maVksNhanDan: z.string().nullish(),
  tenVksNhanDan: z.string().nullish(),
  maDiaDiemBanHanh: z.string().nullish(),
  tenDiaDiemBanHanh: z.string().nullish(),
  maPhanLoaiNguon: z.string().nullish(),
  tenPhanLoaiDonViKienNghi: z.string().nullish(),
  idVuViec: z.string().nullish(),
  maTrangThai: z.string().nullish(),
  idVuAnGoc: z.string().nullish(),
  trangThaiCapSo027: z.string().nullish(),
  qdKhoiTo: z.string().nullish(),
  ngayKhoiTo: z.string().nullish(),
  loaiKhoiTo: z.number().nullish(),
  dauHieuToiPham: z.string().nullish(),
  suDungDiaChiCu: z.boolean().nullish(),
  thuLyChinh: z.boolean().nullish(),
  vuDiaDiemList: z.array(VuDiaDiemResponse_Schema),
  vuDieuLuatList: z.array(VuDieuLuatResponse_Schema),
  vuAnThongKe: VuAnThongKeResponse_Schema
});